requirements.txt
PyJWT
aiohttp
alembic
asyncpg
bcrypt
fastapi
itsdangerous
pydantic
pydantic[email]
python-Levenshtein
psycopg2 # or psycopg2-binary
python-dotenv
python-multipart
sqlalchemy
thefuzz
uvicorn